requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.5.0
tonclient>=1.41.0
